refactor: move files and change to TS extension. Move git  file from lib/buildpacks to lib/git
